fix: Remove dependence on mock data
Before, mock data was displayed on start-up and only updated
when search results came in. This also lets us get rid of the
local Artist class in favor of the one from the API wrapper.
Last, it fixes the bug where we couldn't display more search
results than the length of the mock artists.